minimig: Add conversion between Latin-1 and UTF-8
The Amiga supports files/directories with Latin-1 characters in their
names, so we can support this by converting them to UTF-8 on the Linux
side.

Latin-1 is a subset of UTF-8, so there are no problems converting in
this direction. For UTF-8 to Lati-1, any directories/files created on
the Linux side that are not 100% representable with Latin-1 characters
will be silently ignored and not passed to the Amiga side.